refactor chat local artifact helpers
